requirements.txt
fastapi
uvicorn
pymongo
dnspython
python-multipart
jinja2
passlib[bcrypt]
email-validator
python-jose
reportlab
python-dotenv
llama-parse==0.3.9
google-generativeai==0.3.2
jinja2==3.1.2
python-multipart==0.0.6
email-validator==2.1.0.post1
pydantic<2.0
PyPDF2
docx2txt
pdfminer
pdfminer.six==20240706
python-dateutil
textract